Updated PlayerController and PlayerCamera Scripts
So finally, we are getting rid of the annoying player controller methods that we used in the previous semester and we have a player controller script that relies only on rigidbody. The player camera is also in its own seperate script which gives us more flexibility with the camera script. The code is not commented yet because the code is not done yet, I still have to implement the jumping and dashing. Hopefully those will be done by tomorrow in the afternoon. I also intend to have a white box of the game done by next week.