Player Rigidbody Based Movement, Updated Camera Movement, and GreyBox Map
Firstly, the player is finally ditching the CharacterController Library and relying on the Unity Physics Engine, Rigidbody. The player can walk, run, jump.
Secondly, The camera has its own script which is incharge of the camera movement instead of the camera code being embedded with the player script.
Finally, I have made a simple map to test the player movement. Just be advised, there's a bug with the player prefab where a small portion of the player goes through the floor which affects the jumping a little bit.