Player Movement Fixed (Hopefully)
The player was following the direction of the camera. So if the camera was pointing upwards, the player will fly. So instead of the player following the camera. There's the direction object that will grab values from the camera and the player will just follow the direction object. So there's significant changes. Also I made the player bigger.