Updates Made to the GameManager, MenuManager, Old Main Menu is Disabled
So the GameManager no longer in has the StartMenu State as the game doesn't need it anymore because the Main Menu is now in another scene instead of infront of the player camera.  Same goes for the Menu Manager. Also, I have disabled the old main menu because the player don't have to go through two different main menus when the game is launched.