Background Music Works, Shadowy People Added, Heartbeat added (sort of)
The shadowy people is implemented. but they don't affect the player because there's a bug that I am trying to figure out how to fix.

The background music does adapt the amount of time left on the timer.

Heartbeat is implemented by it doesn't work for some reason.

The player has walking sounds now but I need a new sound effect for walking that is more consistent.